:truck: Move repeating instrument API methods to separate file
Since they are under a separate section in the API playground, i'd prefer to keep them separated here, also because they can export events as well as instruments.
